correct typo in filename and test to check takeaway has method add_to_order
